fix(cuik): relocate CGR featurizer to reaction.py and guard cuik_molmaker>=0.3
Move CuikmolmakerCGRFeaturizer from molecule.py into reaction.py next to
CondensedGraphOfReactionFeaturizer; update package exports and the direct
import in cli/predict.py.

Drop the Featurizer[...] base. The class takes two required lists (reactants
and products), not the single input the Featurizer[S, T] contract models;
batch_reaction_featurizer takes two lists and the dataset calls the featurizer
directly. The isinstance check in cli/predict.py targets the concrete class,
so it is unaffected.

Replace the stale is_cuikmolmaker_available() call, which was removed when the
mandatory-dependency change merged from main and left the featurizer raising
NameError on instantiation, with a hasattr(cuik_molmaker,
"batch_reaction_featurizer") capability check that raises an informative error
when the installed cuik_molmaker predates 0.3. Bump environment.yml to
cuik_molmaker >=0.3 accordingly.
